[TRIVIAL] move metrics endpoints into observe crate
