+++ crates/observe/Cargo.toml
@@ -7,6 +7,7 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 atty = "0.2"
+async-trait = { workspace = true }
 console-subscriber = "0.3.0"
 futures = { workspace = true }
 once_cell = { workspace = true }
@@ -17,6 +18,7 @@ time = { workspace = true }
 tokio = { workspace = true, features = [ "fs" ] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["env-filter", "fmt", "time"] }
+warp = { workspace = true }
 
 [lints]
 workspace = true
