+++ crates/observe/src/metrics.rs
@@ -1,4 +1,10 @@
-use {once_cell::sync::OnceCell, prometheus::Encoder, std::collections::HashMap};
+use {
+    once_cell::sync::OnceCell,
+    prometheus::Encoder,
+    std::{collections::HashMap, convert::Infallible, net::SocketAddr, sync::Arc},
+    tokio::task::{self, JoinHandle},
+    warp::{Filter, Rejection, Reply},
+};
 
 /// Global metrics registry used by all components.
 static REGISTRY: OnceCell<prometheus_metric_storage::StorageRegistry> = OnceCell::new();
@@ -57,4 +63,39 @@ pub fn encode(registry: &prometheus::Registry) -> String {
     let mut buffer = Vec::new();
     encoder.encode(&registry.gather(), &mut buffer).unwrap();
     String::from_utf8(buffer).unwrap()
+}
+
+pub const DEFAULT_METRICS_PORT: u16 = 9586;
+
+#[async_trait::async_trait]
+pub trait LivenessChecking: Send + Sync {
+    async fn is_alive(&self) -> bool;
+}
+
+pub fn serve_metrics(liveness: Arc<dyn LivenessChecking>, address: SocketAddr) -> JoinHandle<()> {
+    let filter = handle_metrics().or(handle_liveness(liveness));
+    tracing::info!(%address, "serving metrics");
+    task::spawn(warp::serve(filter).bind(address))
+}
+
+// `/metrics` route exposing encoded prometheus data to monitoring system
+pub fn handle_metrics() -> impl Filter<Extract = (impl Reply,), Error = Rejection> + Clone {
+    let registry = get_registry();
+    warp::path("metrics").map(move || encode(registry))
+}
+
+fn handle_liveness(
+    liveness_checker: Arc<dyn LivenessChecking>,
+) -> impl Filter<Extract = (impl Reply,), Error = Rejection> + Clone {
+    warp::path("liveness").and_then(move || {
+        let liveness_checker = liveness_checker.clone();
+        async move {
+            let status = if liveness_checker.is_alive().await {
+                warp::http::StatusCode::OK
+            } else {
+                warp::http::StatusCode::SERVICE_UNAVAILABLE
+            };
+            Result::<_, Infallible>::Ok(warp::reply::with_status(warp::reply(), status))
+        }
+    })
 }
