+++ crates/shared/src/lib.rs
@@ -24,7 +24,6 @@ pub mod http_client;
 pub mod http_solver;
 pub mod interaction;
 pub mod maintenance;
-pub mod metrics;
 pub mod order_quoting;
 pub mod order_validation;
 pub mod price_estimation;
